requirements.txt
python-dotenv==1.0.0
aiohttp==3.9.1
httpx==0.25.2
aiortc==1.6.0

# Note: aiortc requires FFmpeg system libraries
# Install system dependencies first, then install aiortc:
#
# Ubuntu/Debian:
#   sudo apt-get update
#   sudo apt-get install libavformat-dev libavcodec-dev libavdevice-dev libavutil-dev libavfilter-dev libswscale-dev libswresample-dev pkg-config
#   pip install aiortc==1.6.0
#
# macOS:
#   brew install ffmpeg pkg-config
#   pip install aiortc==1.6.0
#
# If you don't want to install FFmpeg, you can use the Node.js examples instead.
